requirements.txt
streamlit==1.47.1
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
openai==1.97.1
python-dotenv==1.1.1
PyPDF2==3.0.1
pytesseract==0.3.13
pillow==10.4.0
pandas==2.2.3
numpy==2.1.2
reportlab==4.2.5
